backend/requirements_test.txt
alembic==1.12.1
annotated-types==0.5.0
anyio==3.7.1
asn1crypto==1.5.1
atomicwrites==1.4.1
attrs==22.1.0
bcrypt==4.0.0
boto3==1.35.61
botocore==1.35.99
cachetools==5.5.2
certifi==2022.6.15
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coverage==7.8.0
cryptography==37.0.4
distlib==0.3.9
distro==1.9.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==2.1.0
et_xmlfile==2.0.0
fastapi==0.104.0
filelock==3.18.0
google-auth==2.38.0
google-auth-oauthlib==1.2.2
greenlet==2.0.2
h11==0.13.0
httpcore==0.18.0
httpx==0.25.0
identify==2.6.9
idna==3.3
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==1.1.1
isort==6.0.1
jiter==0.9.0
jmespath==1.0.1
loguru==0.7.3
Mako==1.2.1
MarkupSafe==2.1.1
mirakuru==2.6.0
nodeenv==1.9.1
numpy==2.2.3
oauthlib==3.2.2
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==21.3
pandas==2.2.3
passlib==1.7.4
pg8000==1.31.2
platformdirs==4.3.7
pluggy==1.0.0
port-for==0.7.4
pre_commit==4.2.0
psutil==7.0.0
psycopg2-binary==2.9.7
py==1.11.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==6.1.1
pytest-mock==3.14.0
pytest-postgresql==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.1
PyYAML==6.0.2
requests==2.28.1
requests-oauthlib==2.0.0
rsa==4.9
ruff==0.9.10
s3transfer==0.10.4
scramp==1.4.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==2.0.22
sqlalchemy-cockroachdb==2.0.2
starlette==0.27.0
testing.common.database==2.0.3
testing.postgresql==1.3.0
tomli==2.0.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.12
uvicorn==0.18.2
virtualenv==20.30.0
zipp==3.8.1
